Refactored csvComponents
